+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Colored logs for development and debugging
 - Common test fixtures
 - Some badges on README page
+- Tests for cases and samples endpoints
 ### Fixed
 - Bugs preventing the gunicorn app to launch
 - Code to compose DB url to work when app is invoked from docker-compose
@@ -23,3 +24,4 @@
 - Renamed table `Individuals` table to the more general `Samples`
 - Renamed table `Regions` table to `Intervals`
 - Use uvicorn logging and avoid printing logs twice
+- Modified samples and cases endpoints to interact with database via CRUD utils